Massive Revision
Code is back to compiling and running after a big code revision and
optimisation. Changes include the removal of the TaggedByID interface
and a massive general cleanup of the various 'Object' classes and their
Builders. Lots of optimisations have been made on the collections front
with the use of ArrayLists being cut back to allow for better suited
collections.